+++ src/components/Charts/Bar.vue
