+++ src/components/Charts/Liquid.vue
