+++ src/components/DescriptionList/index.js
@@ -1,0 +1,2 @@
+import DescriptionList from './DescriptionList'
+export default DescriptionList
