+++ src/components/GlobalFooter/index.js
@@ -1,0 +1,2 @@
+import GlobalFooter from './GlobalFooter'
+export default GlobalFooter
