+++ src/components/GlobalHeader/index.js
@@ -1,0 +1,2 @@
+import GlobalHeader from './GlobalHeader'
+export default GlobalHeader
