+++ src/components/NoticeIcon/index.js
@@ -1,0 +1,2 @@
+import NoticeIcon from './NoticeIcon'
+export default NoticeIcon
