+++ src/components/PageHeader/index.js
@@ -1,0 +1,2 @@
+import PageHeader from './PageHeader'
+export default PageHeader
