+++ src/components/SettingDrawer/SettingDrawer.vue
@@ -168,10 +168,10 @@
 </template>
 
 <script>
-import DetailList from '@/components/tools/DetailList'
-import SettingItem from '@/components/setting/SettingItem'
+import { DetailList } from '@/components'
+import SettingItem from './SettingItem'
 import config from '@/config/defaultSettings'
-import { updateTheme, updateColorWeak, colorList } from '@/components/tools/setting'
+import { updateTheme, updateColorWeak, colorList } from './settingConfig'
 import { mixin, mixinDevice } from '@/utils/mixin'
 
 export default {
