+++ src/components/SettingDrawer/index.js
@@ -1,0 +1,2 @@
+import SettingDrawer from './SettingDrawer'
+export default SettingDrawer
