--- src/components/layouts/index.js
@@ -1,7 +1,0 @@
-import UserLayout from '@/components/layouts/UserLayout'
-import BlankLayout from '@/components/layouts/BlankLayout'
-import BasicLayout from '@/components/layouts/BasicLayout'
-import RouteView from '@/components/layouts/RouteView'
-import PageView from '@/components/layouts/PageView'
-
-export { UserLayout, BasicLayout, BlankLayout, RouteView, PageView }
