+++ src/components/tools/Breadcrumb.vue
@@ -1,9 +1,10 @@
 <template>
   <a-breadcrumb class="breadcrumb">
-    <a-breadcrumb-item v-for="(item, index) in breadList" :key="index">
-      <router-link v-if="item.name != name" :to="{ path: item.path }">
-        {{ item.meta.title }}
-      </router-link>
+    <a-breadcrumb-item v-for="(item, index) in breadList" :key="item.name">
+      <router-link
+        v-if="item.name != name && index != 1"
+        :to="{ path: item.path === '' ? '/' : item.path }"
+      >{{ item.meta.title }}</router-link>
       <span v-else>{{ item.meta.title }}</span>
     </a-breadcrumb-item>
   </a-breadcrumb>
@@ -22,14 +23,12 @@ export default {
   },
   methods: {
     getBreadcrumb () {
-      console.log('this.$route.matched', this.$route.matched)
-
       this.breadList = []
-      this.breadList.push({ name: 'index', path: '/dashboard/', meta: { title: '首页' } })
+      // this.breadList.push({name: 'index', path: '/dashboard/', meta: {title: '首页'}})
 
       this.name = this.$route.name
-      this.$route.matched.forEach((item) => {
-        // item.meta.name === 'dashboard' ? item.path = '/dashboard' : this.$route.path === item.path
+      this.$route.matched.forEach(item => {
+        // item.name !== 'index' && this.breadList.push(item)
         this.breadList.push(item)
       })
     }
@@ -43,5 +42,4 @@ export default {
 </script>
 
 <style scoped>
-
 </style>
